requirements.txt
opencv-python
numpy
pandas
pymupdf
Pillow
